Movie_recommender_system/requirements.txt
fastapi
uvicorn
streamlit
pandas
numpy
scikit-learn
scikit-surprise
joblib
requests
cython